Get Stations

Get Stations

Usage

get_stations(
  type = c("list", "data.frame", "crosstable"),
  run_parallel = TRUE,
  n_cores = parallel::detectCores() - 1L,
  debug = TRUE
)

Arguments

type

vector of character describing the type(s) of output(s) to be returned. Expected values (and default): c("list", "data.frame", "crosstable"). If only one value is given the data is returned in the expected type. If more than one values are given, a list is returned with one list element per type.

run_parallel

default: TRUE

n_cores

number of cores to use if run_parallel = TRUE. Default: one less than the detected number of cores.

debug

logical indicating whether or not to show debug messages

Value

list with general station "overview" (either as list "overview_list" or as data.frame "overview_df") and a crosstable with information which parameters is available per station ("x" if available, NA if not)

Examples

stations <- wasserportal::get_stations(n_cores = 2L)
#> Importing 10 station overviews from Wasserportal Berlin ... ok. (9.50 secs) 
str(stations)
#> List of 3
#>  $ overview_list:List of 10
#>   ..$ surface_water.water_level         : tibble [113 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:113] "5865900" "5827103" "5870400" "5865300" ...
#>   .. ..$ Messstellenname  : chr [1:113] "Allee der Kosmonauten" "Allendestraße" "Alsenbrücke" "Am Bahndamm" ...
#>   .. ..$ Gewaesser        : chr [1:113] "M.-H.-Grenzgr." "Müggelspree" "Griebnitzkanal" "Wuhle" ...
#>   .. ..$ Betreiber        : chr [1:113] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:113] "17.05.2026 05:30" "03.03.2026" "16.05.2026" "25.01.2023 07:00" ...
#>   .. ..$ Wasserstand      : chr [1:113] "45.73" "32.30" "29.38" "32.31" ...
#>   .. ..$ Einheit          : chr [1:113] "m+NHN" "m+NHN" "m+NHN" "m+NHN" ...
#>   .. ..$ Ganglinien       : chr [1:113] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865300" ...
#>   .. ..$ Klassifikation   : chr [1:113] "normal" "inaktiv" "niedrig" "inaktiv" ...
#>   .. ..$ stammdaten_link  : chr [1:113] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865300" ...
#>   ..$ surface_water.flow                : tibble [36 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:36] "5865300" "5864801" "5867401" "5867900" ...
#>   .. ..$ Messstellenname  : chr [1:36] "Am Bahndamm" "Am Kienberg" "Bürgerpark" "Eisenbahnbrücke" ...
#>   .. ..$ Gewaesser        : chr [1:36] "Wuhle" "Hellersdorfer Graben" "Panke" "Nordgraben" ...
#>   .. ..$ Betreiber        : chr [1:36] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:36] "24.01.2023 07:00" "17.05.2026 05:30" "17.05.2026 05:30" "16.05.2026" ...
#>   .. ..$ _Durchfluss      : chr [1:36] "0.156" "0.002" "0.692" "1.01" ...
#>   .. ..$ Einheit          : chr [1:36] "m³/s" "m³/s" "m³/s" "m³/s" ...
#>   .. ..$ Ganglinie        : chr [1:36] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5865300" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5864801" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5867401" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5867900" ...
#>   .. ..$ Klassifikation   : chr [1:36] "inaktiv" "keine" "niedrig" "niedrig" ...
#>   .. ..$ stammdaten_link  : chr [1:36] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5865300" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5864801" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5867401" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5867900" ...
#>   ..$ surface_water.temperature         : tibble [71 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:71] "601" "151" "153" "139" ...
#>   .. ..$ Messstellenname  : chr [1:71] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Fischerinsel" ...
#>   .. ..$ Gewaesser        : chr [1:71] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Spree" ...
#>   .. ..$ Betreiber        : chr [1:71] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:71] "17.05.2026 05:00" "17.05.2026 05:00" "27.10.2025 23:45" "17.05.2026 05:00" ...
#>   .. ..$ Wassertemperatur : chr [1:71] "14.79" "14.81" "10.10" "15.10" ...
#>   .. ..$ Einheit          : chr [1:71] "°C" "°C" "°C" "°C" ...
#>   .. ..$ Ganglinie        : chr [1:71] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=139" ...
#>   .. ..$ Klassifikation   : chr [1:71] ">10 - 15°C" ">10 - 15°C" "inaktiv" ">15 - 20°C" ...
#>   .. ..$ stammdaten_link  : chr [1:71] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=139" ...
#>   ..$ surface_water.conductivity        : tibble [18 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:18] "601" "151" "153" "139" ...
#>   .. ..$ Messstellenname  : chr [1:18] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Fischerinsel" ...
#>   .. ..$ Gewaesser        : chr [1:18] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Spree" ...
#>   .. ..$ Betreiber        : chr [1:18] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:18] "17.05.2026 05:00" "17.05.2026 05:00" "27.10.2025 23:45" "17.05.2026 05:00" ...
#>   .. ..$ Leitfaehigkeit   : chr [1:18] "859" "840" "829" "831" ...
#>   .. ..$ Einheit          : chr [1:18] "µS/cm" "µS/cm" "µS/cm" "µS/cm" ...
#>   .. ..$ Ganglinie        : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=139" ...
#>   .. ..$ Klassifikation   : chr [1:18] ">800 - 1000" ">800 - 1000" "inaktiv" ">800 - 1000" ...
#>   .. ..$ stammdaten_link  : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=139" ...
#>   ..$ surface_water.ph                  : tibble [18 × 9] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:18] "601" "151" "153" "139" ...
#>   .. ..$ Messstellenname  : chr [1:18] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Fischerinsel" ...
#>   .. ..$ Gewaesser        : chr [1:18] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Spree" ...
#>   .. ..$ Betreiber        : chr [1:18] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:18] "17.05.2026 05:00" "17.05.2026 05:00" "27.10.2025 23:45" "17.05.2026 05:00" ...
#>   .. ..$ pHWert           : chr [1:18] "7.28" "7.31" "7.57" "7.45" ...
#>   .. ..$ Ganglinie        : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=139" ...
#>   .. ..$ Klassifikation   : chr [1:18] ">7.0 - 7.5" ">7.0 - 7.5" "inaktiv" ">7.0 - 7.5" ...
#>   .. ..$ stammdaten_link  : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=139" ...
#>   ..$ surface_water.oxygen_concentration: tibble [18 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:18] "601" "151" "153" "139" ...
#>   .. ..$ Messstellenname  : chr [1:18] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Fischerinsel" ...
#>   .. ..$ Gewaesser        : chr [1:18] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Spree" ...
#>   .. ..$ Betreiber        : chr [1:18] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:18] "17.05.2026 05:00" "17.05.2026 05:00" "27.10.2025 23:45" "17.05.2026 05:00" ...
#>   .. ..$ Sauerstoffgehalt : chr [1:18] "6.56" "6.07" "8.37" "5.80" ...
#>   .. ..$ Einheit          : chr [1:18] "mg/l" "mg/l" "mg/l" "mg/l" ...
#>   .. ..$ Ganglinie        : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=139" ...
#>   .. ..$ Klassifikation   : chr [1:18] ">5 - 10" ">5 - 10" "inaktiv" ">5 - 10" ...
#>   .. ..$ stammdaten_link  : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=139" ...
#>   ..$ surface_water.oxygen_saturation   : tibble [18 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:18] "601" "151" "153" "139" ...
#>   .. ..$ Messstellenname  : chr [1:18] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Fischerinsel" ...
#>   .. ..$ Gewaesser        : chr [1:18] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Spree" ...
#>   .. ..$ Betreiber        : chr [1:18] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:18] "16.05.2026 00:00" "16.05.2026 00:00" "27.10.2025 23:45" "27.10.2025 23:45" ...
#>   .. ..$ Parameterwert    : chr [1:18] "64.30" "54.83" "74.55" "" ...
#>   .. ..$ Einheit          : chr [1:18] "%" "%" "%" "%" ...
#>   .. ..$ Ganglinie        : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=139" ...
#>   .. ..$ Klassifikation   : chr [1:18] ">50 - 75" ">50 - 75" "inaktiv" "nicht aktuell" ...
#>   .. ..$ stammdaten_link  : chr [1:18] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=139" ...
#>   ..$ surface_water.quality             : tibble [56 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer : chr [1:56] "102" "105" "110" "115" ...
#>   .. ..$ Messstellenname   : chr [1:56] "Gosener Graben - Gosener Landstr." "Dämeritzsee - Seemitte" "Müggelspree - Fähre Rahnsdorf" "Großer Müggelsee - Seemitte" ...
#>   .. ..$ Datum             : chr [1:56] "07.04.2026 09:12" "08.04.2026 10:35" "08.04.2026 10:11" "08.04.2026 09:52" ...
#>   .. ..$ Entnahmetiefe     : chr [1:56] "0.5 m" "0.5 m" "0.5 m" "0.5 m" ...
#>   .. ..$ Parameterwert     : chr [1:56] "9.5" "10.1" "9.1" "8.9" ...
#>   .. ..$ Einheit           : chr [1:56] "°C" "°C" "°C" "°C" ...
#>   .. ..$ _Bestimmungsgrenze: chr [1:56] "-" "-" "-" "-" ...
#>   .. ..$ Ganglinie         : chr [1:56] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=102&nstoffid=448" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=105&nstoffid=448" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=110&nstoffid=448" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=115&nstoffid=448" ...
#>   .. ..$ Klassifikation    : chr [1:56] "keine" "keine" "keine" "keine" ...
#>   .. ..$ stammdaten_link   : chr [1:56] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=102" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=105" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=110" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=115" ...
#>   ..$ groundwater.level                 : tibble [889 × 11] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer        : chr [1:889] "1" "2" "3" "4" ...
#>   .. ..$ Bezirk                   : chr [1:889] "Reinickendorf" "Reinickendorf" "Reinickendorf" "Reinickendorf" ...
#>   .. ..$ Auspraegung              : chr [1:889] "GW-Stand" "GW-Stand" "GW-Stand + GW-Güte" "GW-Stand" ...
#>   .. ..$ Grundwasserleiter        : chr [1:889] "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" ...
#>   .. ..$ Grundwasserspannung      : chr [1:889] "gespannt" "ungespannt" "gespannt" "ungespannt" ...
#>   .. ..$ Datum                    : chr [1:889] "10.04.2026" "10.04.2026" "10.04.2026" "10.04.2026" ...
#>   .. ..$ Grundwasserstand_m_ue_NHN: chr [1:889] "33.56" "35.08" "33.58" "32.44" ...
#>   .. ..$ Flur_abstand_m_u_GOK     : chr [1:889] "keine Angabe" "2.57" "keine Angabe" "7.44" ...
#>   .. ..$ Ganglinie                : chr [1:889] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=1" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=2" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=3" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=4" ...
#>   .. ..$ Klassifikation           : chr [1:889] "extrem niedrig" "sehr niedrig" "sehr niedrig" "sehr niedrig" ...
#>   .. ..$ stammdaten_link          : chr [1:889] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=1" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=2" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=3" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=4" ...
#>   ..$ groundwater.quality               : tibble [208 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:208] "3" "129" "137" "145" ...
#>   .. ..$ Bezirk           : chr [1:208] "Reinickendorf" "Tempelhof-Schöneberg" "Mitte" "Reinickendorf" ...
#>   .. ..$ Auspraegung      : chr [1:208] "GW-Stand + GW-Güte 5" "GW-Stand, tagesaktuell + GW-Güte 3" "GW-Stand, tagesaktuell + GW-Güte 3" "GW-Stand + GW-Güte 5" ...
#>   .. ..$ Grundwasserleiter: chr [1:208] "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" ...
#>   .. ..$ Datum            : chr [1:208] "08.04.2025" "16.06.2025" "07.04.2025" "08.04.2025" ...
#>   .. ..$ Parameterwert    : chr [1:208] "12.40" "10.80" "14.50" "11.50" ...
#>   .. ..$ Einheit          : chr [1:208] "°C" "°C" "°C" "°C" ...
#>   .. ..$ Ganglinie        : chr [1:208] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=3&nstoffid=10" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=129&nstoffid=10" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=137&nstoffid=10" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=145&nstoffid=10" ...
#>   .. ..$ Klassifikation   : chr [1:208] "keine" "keine" "keine" "keine" ...
#>   .. ..$ stammdaten_link  : chr [1:208] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=3" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=129" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=137" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=145" ...
#>  $ overview_df  :Classes ‘data.table’ and 'data.frame':	1445 obs. of  29 variables:
#>   ..$ key                      : chr [1:1445] "surface_water.water_level" "surface_water.water_level" "surface_water.water_level" "surface_water.water_level" ...
#>   ..$ Messstellennummer        : chr [1:1445] "5865900" "5827103" "5870400" "5865300" ...
#>   ..$ Messstellenname          : chr [1:1445] "Allee der Kosmonauten" "Allendestraße" "Alsenbrücke" "Am Bahndamm" ...
#>   ..$ Gewaesser                : chr [1:1445] "M.-H.-Grenzgr." "Müggelspree" "Griebnitzkanal" "Wuhle" ...
#>   ..$ Betreiber                : chr [1:1445] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   ..$ Datum                    : chr [1:1445] "17.05.2026 05:30" "03.03.2026" "16.05.2026" "25.01.2023 07:00" ...
#>   ..$ Wasserstand              : chr [1:1445] "45.73" "32.30" "29.38" "32.31" ...
#>   ..$ Einheit                  : chr [1:1445] "m+NHN" "m+NHN" "m+NHN" "m+NHN" ...
#>   ..$ Ganglinien               : chr [1:1445] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865300" ...
#>   ..$ Klassifikation           : chr [1:1445] "normal" "inaktiv" "niedrig" "inaktiv" ...
#>   ..$ stammdaten_link          : chr [1:1445] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865300" ...
#>   ..$ _Durchfluss              : chr [1:1445] NA NA NA NA ...
#>   ..$ Ganglinie                : chr [1:1445] NA NA NA NA ...
#>   ..$ Wassertemperatur         : chr [1:1445] NA NA NA NA ...
#>   ..$ Leitfaehigkeit           : chr [1:1445] NA NA NA NA ...
#>   ..$ pHWert                   : chr [1:1445] NA NA NA NA ...
#>   ..$ Sauerstoffgehalt         : chr [1:1445] NA NA NA NA ...
#>   ..$ Parameterwert            : chr [1:1445] NA NA NA NA ...
#>   ..$ Entnahmetiefe            : chr [1:1445] NA NA NA NA ...
#>   ..$ _Bestimmungsgrenze       : chr [1:1445] NA NA NA NA ...
#>   ..$ Bezirk                   : chr [1:1445] NA NA NA NA ...
#>   ..$ Auspraegung              : chr [1:1445] NA NA NA NA ...
#>   ..$ Grundwasserleiter        : chr [1:1445] NA NA NA NA ...
#>   ..$ Grundwasserspannung      : chr [1:1445] NA NA NA NA ...
#>   ..$ Grundwasserstand_m_ue_NHN: chr [1:1445] NA NA NA NA ...
#>   ..$ Flur_abstand_m_u_GOK     : chr [1:1445] NA NA NA NA ...
#>   ..$ water_body               : chr [1:1445] "surface_water" "surface_water" "surface_water" "surface_water" ...
#>   ..$ variable                 : chr [1:1445] "water_level" "water_level" "water_level" "water_level" ...
#>   ..$ station_type             : chr [1:1445] "ows" "ows" "ows" "ows" ...
#>   ..- attr(*, ".internal.selfref")=<pointer: 0x15a8114e0> 
#>  $ crosstable   : tibble [1,112 × 12] (S3: tbl_df/tbl/data.frame)
#>   ..$ Messstellennummer: chr [1:1112] "5865900" "5827103" "5870400" "5865300" ...
#>   ..$ Messstellenname  : chr [1:1112] "Allee der Kosmonauten" "Allendestraße" "Alsenbrücke" "Am Bahndamm" ...
#>   ..$ ows              : chr [1:1112] "x" "x" "x" "x" ...
#>   ..$ odf              : chr [1:1112] NA NA NA "x" ...
#>   ..$ owt              : chr [1:1112] "x" NA "x" "x" ...
#>   ..$ olf              : chr [1:1112] NA NA NA NA ...
#>   ..$ oph              : chr [1:1112] NA NA NA NA ...
#>   ..$ oog              : chr [1:1112] NA NA NA NA ...
#>   ..$ oos              : chr [1:1112] NA NA NA NA ...
#>   ..$ opq              : chr [1:1112] NA NA NA NA ...
#>   ..$ gws              : chr [1:1112] NA NA NA NA ...
#>   ..$ gwq              : chr [1:1112] NA NA NA NA ...