Get Stations

Get Stations

Usage

get_stations(
  type = c("list", "data.frame", "crosstable"),
  run_parallel = TRUE,
  n_cores = parallel::detectCores() - 1L,
  debug = TRUE
)

Arguments

type

vector of character describing the type(s) of output(s) to be returned. Expected values (and default): c("list", "data.frame", "crosstable"). If only one value is given the data is returned in the expected type. If more than one values are given, a list is returned with one list element per type.

run_parallel

default: TRUE

n_cores

number of cores to use if run_parallel = TRUE. Default: one less than the detected number of cores.

debug

logical indicating whether or not to show debug messages

Value

list with general station "overview" (either as list "overview_list" or as data.frame "overview_df") and a crosstable with information which parameters is available per station ("x" if available, NA if not)

Examples

stations <- wasserportal::get_stations(n_cores = 2L)
#> Importing 10 station overviews from Wasserportal Berlin ... ok. (9.17 secs) 
str(stations)
#> List of 3
#>  $ overview_list:List of 10
#>   ..$ surface_water.water_level         : tibble [107 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: int [1:107] 5865900 5827103 5870400 5865300 5819900 5864801 5861101 5800107 5800317 5867003 ...
#>   .. ..$ Messstellenname  : chr [1:107] "Allee der Kosmonauten" "Allendestraße" "Alsenbrücke" "Am Bahndamm" ...
#>   .. ..$ Gewaesser        : chr [1:107] "M.-H.-Grenzgr." "Müggelspree" "Griebnitzkanal" "Wuhle" ...
#>   .. ..$ Betreiber        : chr [1:107] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:107] "09.06.2024 05:30" "09.06.2024 05:30" "09.06.2024 05:30" "" ...
#>   .. ..$ Wasserstand      : num [1:107] 6 51 47 NA 83 6 28 69 87 90 ...
#>   .. ..$ Einheit          : chr [1:107] "cm" "cm" "cm" "cm" ...
#>   .. ..$ Ganglinien       : chr [1:107] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865300" ...
#>   .. ..$ Klassifikation   : chr [1:107] "niedrig" "niedrig" "niedrig" "nicht aktuell" ...
#>   .. ..$ stammdaten_link  : chr [1:107] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865300" ...
#>   ..$ surface_water.flow                : tibble [39 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: int [1:39] 5827103 5870400 5865300 5864801 5867401 5867900 5827101 5870100 5826701 5862811 ...
#>   .. ..$ Messstellenname  : chr [1:39] "Allendestraße" "Alsenbrücke" "Am Bahndamm" "Am Kienberg" ...
#>   .. ..$ Gewaesser        : chr [1:39] "Müggelspree" "Griebnitzkanal" "Wuhle" "Hellersdorfer Graben" ...
#>   .. ..$ Betreiber        : chr [1:39] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:39] "" "08.06.2024 00:00" "" "09.06.2024 05:30" ...
#>   .. ..$ Durchfluss       : num [1:39] NA -1.73 NA 0.017 0.603 0.959 2.99 5.95 5.67 6.48 ...
#>   .. ..$ Einheit          : chr [1:39] "m³/s" "m³/s" "m³/s" "m³/s" ...
#>   .. ..$ Ganglinie        : chr [1:39] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5865300" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=odf&station=5864801" ...
#>   .. ..$ Klassifikation   : chr [1:39] "nicht aktuell" "normal" "nicht aktuell" "keine" ...
#>   .. ..$ stammdaten_link  : chr [1:39] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5865300" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=odf&station=5864801" ...
#>   ..$ surface_water.temperature         : tibble [64 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:64] "601" "151" "153" "509" ...
#>   .. ..$ Messstellenname  : chr [1:64] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Landwehrkanal" ...
#>   .. ..$ Gewaesser        : chr [1:64] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Landwehrkanal" ...
#>   .. ..$ Betreiber        : chr [1:64] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:64] "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" ...
#>   .. ..$ Wassertemperatur : chr [1:64] "20.20" "20.47" "20.37" "20.39" ...
#>   .. ..$ Einheit          : chr [1:64] "°C" "°C" "°C" "°C" ...
#>   .. ..$ Ganglinie        : chr [1:64] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=owt&station=509" ...
#>   .. ..$ Klassifikation   : chr [1:64] ">20 - 25°C" ">20 - 25°C" ">20 - 25°C" ">20 - 25°C" ...
#>   .. ..$ stammdaten_link  : chr [1:64] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=owt&station=509" ...
#>   ..$ surface_water.conductivity        : tibble [17 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:17] "601" "151" "153" "509" ...
#>   .. ..$ Messstellenname  : chr [1:17] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Landwehrkanal" ...
#>   .. ..$ Gewaesser        : chr [1:17] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Landwehrkanal" ...
#>   .. ..$ Betreiber        : chr [1:17] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:17] "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" ...
#>   .. ..$ Leitfaehigkeit   : chr [1:17] "846" "807" "804" "800" ...
#>   .. ..$ Einheit          : chr [1:17] "µS/cm" "µS/cm" "µS/cm" "µS/cm" ...
#>   .. ..$ Ganglinie        : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=olf&station=509" ...
#>   .. ..$ Klassifikation   : chr [1:17] ">800 - 1000" ">800 - 1000" ">800 - 1000" ">600 - 800" ...
#>   .. ..$ stammdaten_link  : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=olf&station=509" ...
#>   ..$ surface_water.ph                  : tibble [17 × 9] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:17] "601" "151" "153" "509" ...
#>   .. ..$ Messstellenname  : chr [1:17] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Landwehrkanal" ...
#>   .. ..$ Gewaesser        : chr [1:17] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Landwehrkanal" ...
#>   .. ..$ Betreiber        : chr [1:17] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:17] "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" ...
#>   .. ..$ pHWert           : chr [1:17] "7.56" "7.35" "7.55" "7.59" ...
#>   .. ..$ Ganglinie        : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oph&station=509" ...
#>   .. ..$ Klassifikation   : chr [1:17] ">7.5 - 8.0" ">7.0 - 7.5" ">7.5 - 8.0" ">7.5 - 8.0" ...
#>   .. ..$ stammdaten_link  : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oph&station=509" ...
#>   ..$ surface_water.oxygen_concentration: tibble [17 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:17] "601" "151" "153" "509" ...
#>   .. ..$ Messstellenname  : chr [1:17] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Landwehrkanal" ...
#>   .. ..$ Gewaesser        : chr [1:17] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Landwehrkanal" ...
#>   .. ..$ Betreiber        : chr [1:17] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:17] "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" "09.06.2024 05:00" ...
#>   .. ..$ Sauerstoffgehalt : chr [1:17] "5.76" "5.64" "5.84" "4.10" ...
#>   .. ..$ Einheit          : chr [1:17] "mg/l" "mg/l" "mg/l" "mg/l" ...
#>   .. ..$ Ganglinie        : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oog&station=509" ...
#>   .. ..$ Klassifikation   : chr [1:17] ">5 - 10" ">5 - 10" ">5 - 10" "" ...
#>   .. ..$ stammdaten_link  : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oog&station=509" ...
#>   ..$ surface_water.oxygen_saturation   : tibble [17 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: chr [1:17] "601" "151" "153" "509" ...
#>   .. ..$ Messstellenname  : chr [1:17] "MPS Berlin-Spandauer-Schifffahrtskanal" "MPS Caprivibrücke" "MPS Charlottenburg" "MPS Landwehrkanal" ...
#>   .. ..$ Gewaesser        : chr [1:17] "Berlin-Spandauer-Schifffahrtskanal" "Spree" "Spree" "Landwehrkanal" ...
#>   .. ..$ Betreiber        : chr [1:17] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   .. ..$ Datum            : chr [1:17] "09.06.2024 00:00" "09.06.2024 00:00" "09.06.2024 00:00" "09.06.2024 00:00" ...
#>   .. ..$ Parameterwert    : chr [1:17] "61.94" "66.53" "65.00" "48.10" ...
#>   .. ..$ Einheit          : chr [1:17] "%" "%" "%" "%" ...
#>   .. ..$ Ganglinie        : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=601" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=151" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=153" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=oos&station=509" ...
#>   .. ..$ Klassifikation   : chr [1:17] ">50 - 75" ">50 - 75" ">50 - 75" ">25  - 50" ...
#>   .. ..$ stammdaten_link  : chr [1:17] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=601" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=151" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=153" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=oos&station=509" ...
#>   ..$ surface_water.quality             : tibble [55 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: int [1:55] 102 105 110 115 120 125 130 135 140 150 ...
#>   .. ..$ Messstellenname  : chr [1:55] "Gosener Graben" "Dämeritzsee - Seemitte" "Müggelspree - Fähre Rahnsdorf" "Großer Müggelsee - Seemitte" ...
#>   .. ..$ Datum            : chr [1:55] "08.04.2024 10:16" "14.05.2024 11:07" "14.05.2024 10:47" "14.05.2024 10:12" ...
#>   .. ..$ Entnahmetiefe    : chr [1:55] "0.5 m" "0.5 m" "0.5 m" "0.5 m" ...
#>   .. ..$ Parameterwert    : num [1:55] 16.1 18.1 18.3 17 17.5 18 18.9 18.7 18.6 18.8 ...
#>   .. ..$ Einheit          : chr [1:55] "°C" "°C" "°C" "°C" ...
#>   .. ..$ Bestimmungsgrenze: chr [1:55] "-" "-" "-" "-" ...
#>   .. ..$ Ganglinie        : chr [1:55] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=102&nstoffid=448" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=105&nstoffid=448" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=110&nstoffid=448" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=opq&station=115&nstoffid=448" ...
#>   .. ..$ Klassifikation   : chr [1:55] ">15 -  20" ">15 -  20" ">15 -  20" ">15 -  20" ...
#>   .. ..$ stammdaten_link  : chr [1:55] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=102" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=105" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=110" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=opq&station=115" ...
#>   ..$ groundwater.level                 : tibble [902 × 11] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer        : int [1:902] 1 2 3 4 9 21 24 25 26 30 ...
#>   .. ..$ Bezirk                   : chr [1:902] "Reinickendorf" "Reinickendorf" "Reinickendorf" "Reinickendorf" ...
#>   .. ..$ Auspraegung              : chr [1:902] "GW-Stand" "GW-Stand" "GW-Stand + GW-Güte" "GW-Stand" ...
#>   .. ..$ Grundwasserleiter        : chr [1:902] "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" ...
#>   .. ..$ Grundwasserspannung      : chr [1:902] "gespannt" "ungespannt" "gespannt" "ungespannt" ...
#>   .. ..$ Datum                    : chr [1:902] "03.05.2024" "13.05.2024" "13.05.2024" "13.05.2024" ...
#>   .. ..$ Grundwasserstand_m_ue_NHN: num [1:902] 33.9 35.3 34 32.7 37 ...
#>   .. ..$ Flur_abstand_m_u_GOK     : chr [1:902] "keine Angabe" "2.32" "keine Angabe" "7.14" ...
#>   .. ..$ Ganglinie                : chr [1:902] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=1" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=2" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=3" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gws&station=4" ...
#>   .. ..$ Klassifikation           : chr [1:902] "sehr niedrig" "normal" "hoch" "normal" ...
#>   .. ..$ stammdaten_link          : chr [1:902] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=1" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=2" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=3" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gws&station=4" ...
#>   ..$ groundwater.quality               : tibble [216 × 10] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ Messstellennummer: int [1:216] 3 137 145 149 282 344 358 484 499 580 ...
#>   .. ..$ Bezirk           : chr [1:216] "Reinickendorf" "Mitte" "Reinickendorf" "Mitte" ...
#>   .. ..$ Auspraegung      : chr [1:216] "GW-Stand + GW-Güte 5" "GW-Stand + GW-Güte 5" "GW-Stand + GW-Güte 5" "GW-Stand + GW-Güte 5" ...
#>   .. ..$ Grundwasserleiter: chr [1:216] "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" "Hauptgrundwasserleiter (GWL 1.3 + 2)" ...
#>   .. ..$ Datum            : chr [1:216] "08.11.2023" "15.11.2023" "08.11.2023" "15.11.2023" ...
#>   .. ..$ Parameterwert    : num [1:216] 12.3 15.1 12.7 12.9 12 13.1 12.6 14.2 12.9 13.1 ...
#>   .. ..$ Einheit          : chr [1:216] "°C" "°C" "°C" "°C" ...
#>   .. ..$ Ganglinie        : chr [1:216] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=3&nstoffid=10" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=137&nstoffid=10" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=145&nstoffid=10" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=gwq&station=149&nstoffid=10" ...
#>   .. ..$ Klassifikation   : chr [1:216] "keine" "keine" "keine" "keine" ...
#>   .. ..$ stammdaten_link  : chr [1:216] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=3" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=137" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=145" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=gwq&station=149" ...
#>  $ overview_df  :Classes ‘data.table’ and 'data.frame':	1451 obs. of  29 variables:
#>   ..$ key                      : chr [1:1451] "surface_water.water_level" "surface_water.water_level" "surface_water.water_level" "surface_water.water_level" ...
#>   ..$ Messstellennummer        : chr [1:1451] "5865900" "5827103" "5870400" "5865300" ...
#>   ..$ Messstellenname          : chr [1:1451] "Allee der Kosmonauten" "Allendestraße" "Alsenbrücke" "Am Bahndamm" ...
#>   ..$ Gewaesser                : chr [1:1451] "M.-H.-Grenzgr." "Müggelspree" "Griebnitzkanal" "Wuhle" ...
#>   ..$ Betreiber                : chr [1:1451] "Land Berlin" "Land Berlin" "Land Berlin" "Land Berlin" ...
#>   ..$ Datum                    : chr [1:1451] "09.06.2024 05:30" "09.06.2024 05:30" "09.06.2024 05:30" "" ...
#>   ..$ Wasserstand              : num [1:1451] 6 51 47 NA 83 6 28 69 87 90 ...
#>   ..$ Einheit                  : chr [1:1451] "cm" "cm" "cm" "cm" ...
#>   ..$ Ganglinien               : chr [1:1451] "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=g&thema=ows&station=5865300" ...
#>   ..$ Klassifikation           : chr [1:1451] "niedrig" "niedrig" "niedrig" "nicht aktuell" ...
#>   ..$ stammdaten_link          : chr [1:1451] "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865900" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5827103" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5870400" "https://wasserportal.berlin.de/station.php?anzeige=i&thema=ows&station=5865300" ...
#>   ..$ Durchfluss               : num [1:1451] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ Ganglinie                : chr [1:1451] NA NA NA NA ...
#>   ..$ Wassertemperatur         : chr [1:1451] NA NA NA NA ...
#>   ..$ Leitfaehigkeit           : chr [1:1451] NA NA NA NA ...
#>   ..$ pHWert                   : chr [1:1451] NA NA NA NA ...
#>   ..$ Sauerstoffgehalt         : chr [1:1451] NA NA NA NA ...
#>   ..$ Parameterwert            : chr [1:1451] NA NA NA NA ...
#>   ..$ Entnahmetiefe            : chr [1:1451] NA NA NA NA ...
#>   ..$ Bestimmungsgrenze        : chr [1:1451] NA NA NA NA ...
#>   ..$ Bezirk                   : chr [1:1451] NA NA NA NA ...
#>   ..$ Auspraegung              : chr [1:1451] NA NA NA NA ...
#>   ..$ Grundwasserleiter        : chr [1:1451] NA NA NA NA ...
#>   ..$ Grundwasserspannung      : chr [1:1451] NA NA NA NA ...
#>   ..$ Grundwasserstand_m_ue_NHN: num [1:1451] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ Flur_abstand_m_u_GOK     : chr [1:1451] NA NA NA NA ...
#>   ..$ water_body               : chr [1:1451] "surface_water" "surface_water" "surface_water" "surface_water" ...
#>   ..$ variable                 : chr [1:1451] "water_level" "water_level" "water_level" "water_level" ...
#>   ..$ station_type             : chr [1:1451] "ows" "ows" "ows" "ows" ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ crosstable   : tibble [1,119 × 12] (S3: tbl_df/tbl/data.frame)
#>   ..$ Messstellennummer: chr [1:1119] "5865900" "5827103" "5870400" "5865300" ...
#>   ..$ Messstellenname  : chr [1:1119] "Allee der Kosmonauten" "Allendestraße" "Alsenbrücke" "Am Bahndamm" ...
#>   ..$ ows              : chr [1:1119] "x" "x" "x" "x" ...
#>   ..$ odf              : chr [1:1119] NA "x" "x" "x" ...
#>   ..$ owt              : chr [1:1119] NA NA NA "x" ...
#>   ..$ olf              : chr [1:1119] NA NA NA NA ...
#>   ..$ oph              : chr [1:1119] NA NA NA NA ...
#>   ..$ oog              : chr [1:1119] NA NA NA NA ...
#>   ..$ oos              : chr [1:1119] NA NA NA NA ...
#>   ..$ opq              : chr [1:1119] NA NA NA NA ...
#>   ..$ gws              : chr [1:1119] NA NA NA NA ...
#>   ..$ gwq              : chr [1:1119] NA NA NA NA ...