Download and Read Data from wasserportal.berlin.de

This function downloads and reads CSV files from wasserportal.berlin.de.

read_wasserportal(
  station,
  variables = get_wasserportal_variables(station),
  from_date = as.character(Sys.Date() - 90L),
  type = "single",
  include_raw_time = FALSE
)

Arguments

station
station number, as returned by get_wasserportal_stations

variables
vector of variable identifiers, as returned by get_wasserportal_variables

from_date
Date object (or string in format "yyyy-mm-dd" that can be converted to a Date object representing the first day for which to request data. Default: as.character(Sys.Date() - 90L)

type
one of "single" (the default), "daily", "monthly"

include_raw_time
if TRUE the original time column and the column with the corrected winter time are included in the output. The default is FALSE.

Value

data frame read from the CSV file that the download provides. IMPORTANT: It is not yet clear how to interpret the timestamp, see example

Details

The original timestamps (column timestamps_raw in the example below) are not all plausible, e.g. "31.03.2019 03:00" appears twice! They are corrected (column timestamp_corr) to represent a plausible sequence of timestamps in Berlin Normal Time (UTC+01) Finally, a valid POSIXct timestamp in timezone "Berlin/Europe" (UTC+01 in winter, UTC+02 in summer) is created, together with the additional information on the UTC offset (column UTCOffset, 1 in winter, 2 in summer).

Examples

# Get a list of available water quality stations and variables
stations <- wasserportal::get_wasserportal_stations()
variables <- wasserportal::get_wasserportal_variables()

# Set the start date
from_date <- "2020-03-01"

# Read the timeseries (multiple variables for one station)
water_quality <- wasserportal::read_wasserportal(
  station = stations$MPS_Charlottenburg,
  variables = c(variables["Sauerstoffgehalt"], variables["Leitfaehigkeit"]),
  from_date = from_date, include_raw_time = TRUE
)

#> Reading 'Sauerstoffgehalt' for station 153 (MPS_Charlottenburg) ... ok. (2.45s) 
#> Reading 'Leitfaehigkeit' for station 153 (MPS_Charlottenburg) ... ok. (1.13s) 

# Look at the first few records
head(water_quality)

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 12962 13.03.2020 03:00 13.03.2020 03:00 2020-03-13 03:00:00         1
#> 12963 13.03.2020 03:15 13.03.2020 03:15 2020-03-13 03:15:00         1
#> 12964 13.03.2020 03:30 13.03.2020 03:30 2020-03-13 03:30:00         1
#> 12965 13.03.2020 03:45 13.03.2020 03:45 2020-03-13 03:45:00         1
#> 12966 13.03.2020 04:00 13.03.2020 04:00 2020-03-13 04:00:00         1
#> 12967 13.03.2020 04:15 13.03.2020 04:15 2020-03-13 04:15:00         1
#>       Sauerstoffgehalt Leitfaehigkeit
#> 12962             -777           -777
#> 12963             -777           -777
#> 12964             -777           -777
#> 12965             -777           -777
#> 12966             -777           -777
#> 12967             -777           -777

# Check the metadata
#kwb.utils::getAttribute(water_quality, "metadata")

# Set missing values to NA
water_quality[water_quality == -777] <- NA

# Look at the first few records again
head(water_quality)

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 12962 13.03.2020 03:00 13.03.2020 03:00 2020-03-13 03:00:00         1
#> 12963 13.03.2020 03:15 13.03.2020 03:15 2020-03-13 03:15:00         1
#> 12964 13.03.2020 03:30 13.03.2020 03:30 2020-03-13 03:30:00         1
#> 12965 13.03.2020 03:45 13.03.2020 03:45 2020-03-13 03:45:00         1
#> 12966 13.03.2020 04:00 13.03.2020 04:00 2020-03-13 04:00:00         1
#> 12967 13.03.2020 04:15 13.03.2020 04:15 2020-03-13 04:15:00         1
#>       Sauerstoffgehalt Leitfaehigkeit
#> 12962               NA             NA
#> 12963               NA             NA
#> 12964               NA             NA
#> 12965               NA             NA
#> 12966               NA             NA
#> 12967               NA             NA

### How was the original timestamp interpreted?

# Determine the days at which summer time starts and ends, respectively
from_year <- as.integer(substr(from_date, 1L, 4L))
switches <- kwb.datetime::date_range_CEST(from_year)

# Reformat to dd.mm.yyyy
switches <- kwb.datetime::reformatTimestamp(switches, "%Y-%m-%d", "%d.%m.%Y")

# Define a pattern to look for timestamps "around" the switches
pattern <- paste(switches, "0[1-4]", collapse = "|")

# Look at the data for these timestamps
water_quality[grepl(pattern, water_quality$timestamp_raw), ]

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 31290 29.03.2020 01:00 29.03.2020 01:00 2020-03-29 01:00:00         1
#> 31291 29.03.2020 01:15 29.03.2020 01:15 2020-03-29 01:15:00         1
#> 31292 29.03.2020 01:30 29.03.2020 01:30 2020-03-29 01:30:00         1
#> 31293 29.03.2020 01:45 29.03.2020 01:45 2020-03-29 01:45:00         1
#> 31294 29.03.2020 02:00 29.03.2020 02:00 2020-03-29 03:00:00         2
#> 31295 29.03.2020 02:15 29.03.2020 02:15 2020-03-29 03:15:00         2
#> 31296 29.03.2020 02:30 29.03.2020 02:30 2020-03-29 03:30:00         2
#> 31297 29.03.2020 02:45 29.03.2020 02:45 2020-03-29 03:45:00         2
#> 31298 29.03.2020 03:00 29.03.2020 03:00 2020-03-29 04:00:00         2
#> 31299 29.03.2020 03:15 29.03.2020 03:15 2020-03-29 04:15:00         2
#> 31300 29.03.2020 03:30 29.03.2020 03:30 2020-03-29 04:30:00         2
#> 31301 29.03.2020 03:45 29.03.2020 03:45 2020-03-29 04:45:00         2
#> 31302 29.03.2020 04:00 29.03.2020 04:00 2020-03-29 05:00:00         2
#> 31303 29.03.2020 04:15 29.03.2020 04:15 2020-03-29 05:15:00         2
#> 31304 29.03.2020 04:30 29.03.2020 04:30 2020-03-29 05:30:00         2
#> 31305 29.03.2020 04:45 29.03.2020 04:45 2020-03-29 05:45:00         2
#> 27450 25.10.2020 01:00 25.10.2020 01:00 2020-10-25 02:00:00         2
#> 27451 25.10.2020 01:15 25.10.2020 01:15 2020-10-25 02:15:00         2
#> 27452 25.10.2020 01:30 25.10.2020 01:30 2020-10-25 02:30:00         2
#> 27453 25.10.2020 01:45 25.10.2020 01:45 2020-10-25 02:45:00         2
#> 27454 25.10.2020 02:00 25.10.2020 02:00 2020-10-25 02:00:00         1
#> 27455 25.10.2020 02:15 25.10.2020 02:15 2020-10-25 02:15:00         1
#> 27456 25.10.2020 02:30 25.10.2020 02:30 2020-10-25 02:30:00         1
#> 27457 25.10.2020 02:45 25.10.2020 02:45 2020-10-25 02:45:00         1
#> 27458 25.10.2020 03:00 25.10.2020 03:00 2020-10-25 03:00:00         1
#> 27459 25.10.2020 03:15 25.10.2020 03:15 2020-10-25 03:15:00         1
#> 27460 25.10.2020 03:30 25.10.2020 03:30 2020-10-25 03:30:00         1
#> 27461 25.10.2020 03:45 25.10.2020 03:45 2020-10-25 03:45:00         1
#> 27462 25.10.2020 04:00 25.10.2020 04:00 2020-10-25 04:00:00         1
#> 27463 25.10.2020 04:15 25.10.2020 04:15 2020-10-25 04:15:00         1
#> 27464 25.10.2020 04:30 25.10.2020 04:30 2020-10-25 04:30:00         1
#> 27465 25.10.2020 04:45 25.10.2020 04:45 2020-10-25 04:45:00         1
#>       Sauerstoffgehalt Leitfaehigkeit
#> 31290               NA             NA
#> 31291               NA             NA
#> 31292               NA             NA
#> 31293               NA             NA
#> 31294               NA             NA
#> 31295               NA             NA
#> 31296               NA             NA
#> 31297               NA             NA
#> 31298               NA             NA
#> 31299               NA             NA
#> 31300               NA             NA
#> 31301               NA             NA
#> 31302               NA             NA
#> 31303               NA             NA
#> 31304               NA             NA
#> 31305               NA             NA
#> 27450               NA             NA
#> 27451               NA             NA
#> 27452               NA             NA
#> 27453               NA             NA
#> 27454               NA             NA
#> 27455               NA             NA
#> 27456               NA             NA
#> 27457               NA             NA
#> 27458               NA             NA
#> 27459               NA             NA
#> 27460               NA             NA
#> 27461               NA             NA
#> 27462               NA             NA
#> 27463               NA             NA
#> 27464               NA             NA
#> 27465               NA             NA

# The original timestamps (timestamps_raw) were not all plausible, e.g.
# for March 2019. This seems to have been fixed by the "wasserportal"!
sum(water_quality$timestamp_raw != water_quality$timestamp_corr)

#> [1] 0