Download and Read Data from wasserportal.berlin.de

This function downloads and reads CSV files from wasserportal.berlin.de.

read_wasserportal(
  station,
  variables = get_wasserportal_variables(station),
  from_date = as.character(Sys.Date() - 90L),
  type = "single",
  include_raw_time = FALSE
)

Arguments

station
station number, as returned by get_wasserportal_stations

variables
vector of variable identifiers, as returned by get_wasserportal_variables

from_date
Date object (or string in format "yyyy-mm-dd" that can be converted to a Date object representing the first day for which to request data. Default: as.character(Sys.Date() - 90L)

type
one of "single" (the default), "daily", "monthly"

include_raw_time
if TRUE the original time column and the column with the corrected winter time are included in the output. The default is FALSE.

Value

data frame read from the CSV file that the download provides. IMPORTANT: It is not yet clear how to interpret the timestamp, see example

Details

The original timestamps (column timestamps_raw in the example below) are not all plausible, e.g. "31.03.2019 03:00" appears twice! They are corrected (column timestamp_corr) to represent a plausible sequence of timestamps in Berlin Normal Time (UTC+01) Finally, a valid POSIXct timestamp in timezone "Berlin/Europe" (UTC+01 in winter, UTC+02 in summer) is created, together with the additional information on the UTC offset (column UTCOffset, 1 in winter, 2 in summer).

Examples

# Get a list of available water quality stations and variables
stations <- wasserportal::get_wasserportal_stations()
variables <- wasserportal::get_wasserportal_variables()

# Set the start date
from_date <- "2020-03-01"

# Read the timeseries (multiple variables for one station)
water_quality <- wasserportal::read_wasserportal(
  station = stations$MPS_Charlottenburg,
  variables = c(variables["Sauerstoffgehalt"], variables["Leitfaehigkeit"]),
  from_date = from_date, include_raw_time = TRUE
)

#> Reading 'Sauerstoffgehalt' for station 153 (MPS_Charlottenburg) ... ok. (1.96s) 
#> Reading 'Leitfaehigkeit' for station 153 (MPS_Charlottenburg) ... ok. (1.12s) 

# Look at the first few records
head(water_quality)

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 10658 13.05.2020 02:00 13.05.2020 02:00 2020-05-13 03:00:00         2
#> 10659 13.05.2020 02:15 13.05.2020 02:15 2020-05-13 03:15:00         2
#> 10660 13.05.2020 02:30 13.05.2020 02:30 2020-05-13 03:30:00         2
#> 10661 13.05.2020 02:45 13.05.2020 02:45 2020-05-13 03:45:00         2
#> 10662 13.05.2020 03:00 13.05.2020 03:00 2020-05-13 04:00:00         2
#> 10663 13.05.2020 03:15 13.05.2020 03:15 2020-05-13 04:15:00         2
#>       Sauerstoffgehalt Leitfaehigkeit
#> 10658             7.38            813
#> 10659             7.20            813
#> 10660             7.36            813
#> 10661             7.39            812
#> 10662             7.35            811
#> 10663             7.32            811

# Check the metadata
#kwb.utils::getAttribute(water_quality, "metadata")

# Set missing values to NA
water_quality[water_quality == -777] <- NA

# Look at the first few records again
head(water_quality)

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 10658 13.05.2020 02:00 13.05.2020 02:00 2020-05-13 03:00:00         2
#> 10659 13.05.2020 02:15 13.05.2020 02:15 2020-05-13 03:15:00         2
#> 10660 13.05.2020 02:30 13.05.2020 02:30 2020-05-13 03:30:00         2
#> 10661 13.05.2020 02:45 13.05.2020 02:45 2020-05-13 03:45:00         2
#> 10662 13.05.2020 03:00 13.05.2020 03:00 2020-05-13 04:00:00         2
#> 10663 13.05.2020 03:15 13.05.2020 03:15 2020-05-13 04:15:00         2
#>       Sauerstoffgehalt Leitfaehigkeit
#> 10658             7.38            813
#> 10659             7.20            813
#> 10660             7.36            813
#> 10661             7.39            812
#> 10662             7.35            811
#> 10663             7.32            811

### How was the original timestamp interpreted?

# Determine the days at which summer time starts and ends, respectively
from_year <- as.integer(substr(from_date, 1L, 4L))
switches <- kwb.datetime::date_range_CEST(from_year)

# Reformat to dd.mm.yyyy
switches <- kwb.datetime::reformatTimestamp(switches, "%Y-%m-%d", "%d.%m.%Y")

# Define a pattern to look for timestamps "around" the switches
pattern <- paste(switches, "0[1-4]", collapse = "|")

# Look at the data for these timestamps
water_quality[grepl(pattern, water_quality$timestamp_raw), ]

#>          timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 22654 25.10.2020 01:00 25.10.2020 01:00 2020-10-25 02:00:00         2
#> 22655 25.10.2020 01:15 25.10.2020 01:15 2020-10-25 02:15:00         2
#> 22656 25.10.2020 01:30 25.10.2020 01:30 2020-10-25 02:30:00         2
#> 22657 25.10.2020 01:45 25.10.2020 01:45 2020-10-25 02:45:00         2
#> 22658 25.10.2020 02:00 25.10.2020 02:00 2020-10-25 02:00:00         1
#> 22659 25.10.2020 02:15 25.10.2020 02:15 2020-10-25 02:15:00         1
#> 22660 25.10.2020 02:30 25.10.2020 02:30 2020-10-25 02:30:00         1
#> 22661 25.10.2020 02:45 25.10.2020 02:45 2020-10-25 02:45:00         1
#> 22662 25.10.2020 03:00 25.10.2020 03:00 2020-10-25 03:00:00         1
#> 22663 25.10.2020 03:15 25.10.2020 03:15 2020-10-25 03:15:00         1
#> 22664 25.10.2020 03:30 25.10.2020 03:30 2020-10-25 03:30:00         1
#> 22665 25.10.2020 03:45 25.10.2020 03:45 2020-10-25 03:45:00         1
#> 22666 25.10.2020 04:00 25.10.2020 04:00 2020-10-25 04:00:00         1
#> 22667 25.10.2020 04:15 25.10.2020 04:15 2020-10-25 04:15:00         1
#> 22668 25.10.2020 04:30 25.10.2020 04:30 2020-10-25 04:30:00         1
#> 22669 25.10.2020 04:45 25.10.2020 04:45 2020-10-25 04:45:00         1
#>       Sauerstoffgehalt Leitfaehigkeit
#> 22654               NA             NA
#> 22655               NA             NA
#> 22656               NA             NA
#> 22657               NA             NA
#> 22658               NA             NA
#> 22659               NA             NA
#> 22660               NA             NA
#> 22661               NA             NA
#> 22662               NA             NA
#> 22663               NA             NA
#> 22664               NA             NA
#> 22665               NA             NA
#> 22666               NA             NA
#> 22667               NA             NA
#> 22668               NA             NA
#> 22669               NA             NA

# The original timestamps (timestamps_raw) were not all plausible, e.g.
# for March 2019. This seems to have been fixed by the "wasserportal"!
sum(water_quality$timestamp_raw != water_quality$timestamp_corr)

#> [1] 0