Download and Read Data from wasserportal.berlin.de

This function downloads and reads CSV files from wasserportal.berlin.de.

read_wasserportal(
  station,
  variables = get_wasserportal_variables(station),
  from_date = as.character(Sys.Date() - 90L),
  type = "single",
  include_raw_time = FALSE
)

Arguments

station

station number, as returned by get_wasserportal_stations

variables

vector of variable identifiers, as returned by get_wasserportal_variables

from_date

Date object (or string in format "yyyy-mm-dd" that can be converted to a Date object representing the first day for which to request data. Default: as.character(Sys.Date() - 90L)

type

one of "single" (the default), "daily", "monthly"

include_raw_time

if TRUE the original time column and the column with the corrected winter time are included in the output. The default is FALSE.

Value

data frame read from the CSV file that the download provides. IMPORTANT: It is not yet clear how to interpret the timestamp, see example

Details

The original timestamps (column timestamps_raw in the example below) are not all plausible, e.g. "31.03.2019 03:00" appears twice! They are corrected (column timestamp_corr) to represent a plausible sequence of timestamps in Berlin Normal Time (UTC+01) Finally, a valid POSIXct timestamp in timezone "Berlin/Europe" (UTC+01 in winter, UTC+02 in summer) is created, together with the additional information on the UTC offset (column UTCOffset, 1 in winter, 2 in summer).

Examples

# Get a list of available water quality stations and variables
stations <- wasserportal::get_wasserportal_stations()
variables <- wasserportal::get_wasserportal_variables()

# Set the start date
from_date <- "2020-03-01"

# Read the timeseries (multiple variables for one station)
water_quality <- wasserportal::read_wasserportal(
  station = stations$MPS_Charlottenburg,
  variables = c(variables["Sauerstoffgehalt"], variables["Leitfaehigkeit"]),
  from_date = from_date, include_raw_time = TRUE
)
#> Reading 'Sauerstoffgehalt' for station 153 (MPS_Charlottenburg) ... ok. (1.12s) 
#> Reading 'Leitfaehigkeit' for station 153 (MPS_Charlottenburg) ... ok. (0.54s) 

# Look at the first few records
head(water_quality)
#>         timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 4898 25.12.2020 03:00 25.12.2020 03:00 2020-12-25 03:00:00         1
#> 4899 25.12.2020 03:15 25.12.2020 03:15 2020-12-25 03:15:00         1
#> 4900 25.12.2020 03:30 25.12.2020 03:30 2020-12-25 03:30:00         1
#> 4901 25.12.2020 03:45 25.12.2020 03:45 2020-12-25 03:45:00         1
#> 4902 25.12.2020 04:00 25.12.2020 04:00 2020-12-25 04:00:00         1
#> 4903 25.12.2020 04:15 25.12.2020 04:15 2020-12-25 04:15:00         1
#>      Sauerstoffgehalt Leitfaehigkeit
#> 4898             -777           -777
#> 4899             -777           -777
#> 4900             -777           -777
#> 4901             -777           -777
#> 4902             -777           -777
#> 4903             -777           -777

# Check the metadata
#kwb.utils::getAttribute(water_quality, "metadata")

# Set missing values to NA
water_quality[water_quality == -777] <- NA

# Look at the first few records again
head(water_quality)
#>         timestamp_raw   timestamp_corr       LocalDateTime UTCOffset
#> 4898 25.12.2020 03:00 25.12.2020 03:00 2020-12-25 03:00:00         1
#> 4899 25.12.2020 03:15 25.12.2020 03:15 2020-12-25 03:15:00         1
#> 4900 25.12.2020 03:30 25.12.2020 03:30 2020-12-25 03:30:00         1
#> 4901 25.12.2020 03:45 25.12.2020 03:45 2020-12-25 03:45:00         1
#> 4902 25.12.2020 04:00 25.12.2020 04:00 2020-12-25 04:00:00         1
#> 4903 25.12.2020 04:15 25.12.2020 04:15 2020-12-25 04:15:00         1
#>      Sauerstoffgehalt Leitfaehigkeit
#> 4898               NA             NA
#> 4899               NA             NA
#> 4900               NA             NA
#> 4901               NA             NA
#> 4902               NA             NA
#> 4903               NA             NA

### How was the original timestamp interpreted?

# Determine the days at which summer time starts and ends, respectively
from_year <- as.integer(substr(from_date, 1L, 4L))
switches <- kwb.datetime::date_range_CEST(from_year)

# Reformat to dd.mm.yyyy
switches <- kwb.datetime::reformatTimestamp(switches, "%Y-%m-%d", "%d.%m.%Y")

# Define a pattern to look for timestamps "around" the switches
pattern <- paste(switches, "0[1-4]", collapse = "|")

# Look at the data for these timestamps
water_quality[grepl(pattern, water_quality$timestamp_raw), ]
#> [1] timestamp_raw    timestamp_corr   LocalDateTime    UTCOffset       
#> [5] Sauerstoffgehalt Leitfaehigkeit  
#> <0 rows> (or 0-length row.names)

# The original timestamps (timestamps_raw) were not all plausible, e.g.
# for March 2019. This seems to have been fixed by the "wasserportal"!
sum(water_quality$timestamp_raw != water_quality$timestamp_corr)
#> [1] 0